+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [Enhancement] Support `:earliest` and `:latest` in `Karafka::Admin#seek_consumer_group`.
 - [Maintenance] Use `base64` from RubyGems as it will no longer be part of standard library in Ruby 3.4.
 - [Fix] Support migrating via aliases and plan with aliases usage.
+- [Fix] Active with default set to `false` cannot be overwritten
 
 ## 2.4.0 (2024-04-26)